fix(std/node): "events" and "util" modules (#7170)
